requirements.txt
beautifulsoup4==4.11.1
Django==4.1.5
djangorestframework==3.13.1
lxml==4.9.2
pyTelegramBotAPI==4.9.0
python-dotenv==0.21.0
psycopg2-binary==2.9.3
redis==4.4.2
requests==2.28.2
translate==3.6.1